1124 - Rename Datadog Listener to MetricsListener for consistency. (#1341)
* Renamed Datadog::Listener to MetricsListener.

* Updated CHANGELOG.

* Alias previous Listener class to MetricsListener.

* Added specs for old Listener class.
